refactor deploy routines, update processing logic
